fixed NewEventForm.jsx submit error
