+++ src/Components/Attendees.jsx
@@ -10,7 +10,7 @@ export default function Attendees({ showAttendees, attendees, event, updateEvent
       {showAttendees ? (
       <div key={event.id} className="attendees">
         {attendees.map(attendee => (
-        // add attendee prop here where HTML code used to be
+        // add attendee prop here where HTML code used to be (HTML placed in Attendee component)
         <Attendee key={attendee.id} attendee={attendee} updateEventAttendance={updateEventAttendance} />
         ))}
       </div>
